[Fix] canSendTempPassword nullException 수정
