+++ back/src/main/java/travelRepo/domain/account/service/AccountService.java
@@ -6,8 +6,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.thymeleaf.TemplateEngine;
-import org.thymeleaf.context.Context;
 import travelRepo.domain.account.dto.*;
 import travelRepo.domain.account.entity.Account;
 import travelRepo.domain.account.repository.AccountRepository;
@@ -20,8 +18,6 @@ import travelRepo.domain.likes.likesRepository.LikesRepository;
 import travelRepo.global.common.dto.IdDto;
 import travelRepo.global.exception.BusinessLogicException;
 import travelRepo.global.exception.ExceptionCode;
-import travelRepo.global.mail.EmailMessageDto;
-import travelRepo.global.mail.EmailService;
 import travelRepo.global.upload.service.UploadService;
 
 import java.util.List;
@@ -40,8 +36,6 @@ public class AccountService {
     private final LikesRepository likesRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
     private final UploadService uploadService;
-    private final EmailService emailService;
-    private final TemplateEngine templateEngine;
 
     @Transactional
     public IdDto addAccount(AccountAddReq accountAddReq) {
@@ -115,33 +109,6 @@ public class AccountService {
             }
             return followAccountDetailsRes;
         });
-    }
-
-    @Transactional
-    public void sendTempPasswordGuide(TempPasswordGuideSendReq tempPasswordGuideSendReq) {
-
-        Account account = accountRepository.findByEmail(tempPasswordGuideSendReq.getEmail())
-                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.NOT_FOUND_ACCOUNT));
-
-        account.createTempPassword();
-
-        Context context = new Context();
-        context.setVariable("link", "/accounts/tempPassword/" + account.getId());
-        context.setVariable("nickname", account.getNickname());
-        context.setVariable("linkName", "임시 비밀번호 발급");
-        context.setVariable("message", "임시 비밀번호를 발급하려면 링크를 클릭하세요");
-        context.setVariable("tempPassword", account.getTempPassword());
-        context.setVariable("host", "http://localhost:8080");
-
-        String message = templateEngine.process("mail/simple-link", context);
-
-        EmailMessageDto emailMessageDto = EmailMessageDto.builder()
-                .to(account.getEmail())
-                .subject("TravelRepo, 임시 비밀번호 발급")
-                .message(message)
-                .build();
-
-        emailService.sendEmail(emailMessageDto);
     }
 
     private void verifyDuplicateEmail(AccountAddReq accountAddReq) {
